pyproject.toml
[tool.poetry]
name = "tiktok_watermarkless_api"
version = "0.1.0"
description = ""
authors = ["rdfsx <m315498@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
yt-dlp = {git = "https://github.com/yt-dlp/yt-dlp.git"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
